+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.1] - 2021-04-30
+### Changed
+- Update Bayer strength recommendations for color images
+
+
 ## [2.1.0] - 2021-04-29
 ### Added
 - JSON tags for `OrdereredDitherMatrix`